ADO#1014 build fix 4: CoverageBlock Razor syntax (% expression, string literals in onclick), AlertService decimal/double mismatch

+++ famos/src/FamOs.Web/Services/AlertService.cs
@@ -4,7 +4,7 @@ namespace FamOs.Web.Services;
 
 public class AlertService : IAlertService
 {
-    private const double PremiumDifferentialThreshold = 0.20;
+    private const decimal PremiumDifferentialThreshold = 0.20m;
 
     public List<AlertDto> EvaluateAlerts(
         ComparisonContextDto context,
